no deleting old messages

+++ bot.py
@@ -28,7 +28,7 @@ async def on_message(message):
         data["leaderboard_message_channel_id"] = message.channel.id
         files.write(f"data/{message.guild.id}.pkl", data)
     if message.author.bot:
-        if str(message.channel) == "leaderboard" or str(message.channel) == "counting":
+        if (str(message.channel) == "leaderboard" or str(message.channel) == "counting") and message.author != bot.user:
             await message.delete()
         return
